+++ common/src/main/java/com/coolerpromc/uncrafteverything/screen/custom/UEConfigScreen.java
@@ -41,10 +41,12 @@ public class UEConfigScreen extends AbstractScrollableScreen {
     private boolean outputEnchantedBook = config.outputEnchantedBook();
     private boolean prioritizeVanillaIngredientRecipe = config.prioritizeVanillaIngredientRecipe();
     private boolean restrictAmbiguouslyCraftedItems = config.restrictAmbiguouslyCraftedItems();
+    private boolean allowDamagedNonRepairable = config.allowDamagedNonRepairable();
+    private double minimumDurability = config.minimumDurability();
 
     private Button restrictionTypeButton;
     private Button toggleEnchantedBtn;
-    private Button toggleEnchantmentTypeBtn;
+    private Button toggleExpTypeBtn;
     private Button toggleRestrictAmbiguouslyCraftedItems;
     private Button toggleAllowUnsmithing;
     private Button toggleAllowDamaged;
@@ -53,47 +55,43 @@ public class UEConfigScreen extends AbstractScrollableScreen {
     private Button toggleOnlyAllowDefinedProgression;
     private Button togglePreventModdedIngredientsFromVanillaItems;
     private Button togglePrioritizeVanillaIngredientRecipe;
+    private Button toggleAllowDamagedNonRepairable;
     private MultiLineEditBox restrictionsInput;
     private EditBox experienceInput;
     private MultiLineEditBox restrictedModInput;
+    private EditBox minimumDurabilityInput;
     private Button saveButton;
 
     public UEConfigScreen(Component title) {
-        super(title, 463);
+        super(title, 713);
     }
 
     @Override
     protected void init() {
         int x = this.width / 2 + 10;
         int widgetWidth = this.width - x - 10;
-        int baseY = 30; // Start position, accounting for title and scroll
+        int y = 55;
 
-        // Restriction Type Config
-        restrictionTypeButton =  Button.builder(Component.translatable("screen.uncrafteverything.config.restriction_type_" + restrictionType.toString().toLowerCase()), this::pressRestrictionTypeButton).bounds(x, (int) (baseY - scrollAmount), widgetWidth, 20).build();
+        // Restrictions
+        restrictionTypeButton = Button.builder(Component.translatable("screen.uncrafteverything.config.restriction_type_" + restrictionType.toString().toLowerCase()), this::pressRestrictionTypeButton).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(restrictionTypeButton);
+        y += 25;
 
-        // Restrictions input box
         String joined = String.join("\n", restrictions);
-        restrictionsInput = MultiLineEditBox.builder().setX(x).setY((int) (baseY + 25 - scrollAmount)).build(this.font, widgetWidth, 90, Component.translatable("screen.uncrafteverything.blank"));
+        restrictionsInput = MultiLineEditBox.builder().setX(x).setY((int) (y - scrollAmount)).build(this.font, widgetWidth, 90, Component.translatable("screen.uncrafteverything.blank"));
         restrictionsInput.setValue(joined);
         this.addRenderableWidget(restrictionsInput);
+        y += 95;
 
-        // Toggle for restrictAmbiguouslyCraftedItems
         toggleRestrictAmbiguouslyCraftedItems = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.restrict_ambiguously_", restrictAmbiguouslyCraftedItems)), btn -> {
             restrictAmbiguouslyCraftedItems = !restrictAmbiguouslyCraftedItems;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.restrict_ambiguously_", restrictAmbiguouslyCraftedItems)));
-        }).bounds(x, (int) (baseY + 120 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(toggleRestrictAmbiguouslyCraftedItems);
+        y += 50;
 
-        // Toggle for allowEnchantedItems
-       toggleEnchantedBtn = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.allow_enchanted_", allowEnchantedItems)), btn -> {
-            allowEnchantedItems = !allowEnchantedItems;
-            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.allow_enchanted_", allowEnchantedItems)));
-       }).bounds(x, (int) (baseY + 145 - scrollAmount), widgetWidth, 20).build();
-        this.addRenderableWidget(toggleEnchantedBtn);
-
-        // Toggle for experienceType
-        toggleEnchantmentTypeBtn = Button.builder(
+        // Experience
+        toggleExpTypeBtn = Button.builder(
                 Component.translatable("screen.uncrafteverything.config.exp_type_" + experienceType.toString().toLowerCase()),
                 btn -> {
                     UncraftEverythingConfig.ExperienceType[] values = UncraftEverythingConfig.ExperienceType.values();
@@ -101,68 +99,94 @@ public class UEConfigScreen extends AbstractScrollableScreen {
                     experienceType = next;
                     btn.setMessage(Component.translatable("screen.uncrafteverything.config.exp_type_" + next.toString().toLowerCase()));
                 }
-        ).bounds(x, (int) (baseY + 170 - scrollAmount), widgetWidth, 20).build();
-        this.addRenderableWidget(toggleEnchantmentTypeBtn);
+        ).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(toggleExpTypeBtn);
+        y += 25;
 
-        // Experience input box
-        experienceInput = new EditBox(this.font, x, (int) (baseY + 195 - scrollAmount), widgetWidth, 20,
-                Component.translatable("screen.uncrafteverything.blank"));
+        experienceInput = new EditBox(this.font, x, (int) (y - scrollAmount), widgetWidth, 20, Component.translatable("screen.uncrafteverything.blank"));
         experienceInput.setValue(Integer.toString(experience));
         this.addRenderableWidget(experienceInput);
+        y += 50;
 
-        // Toggle for allowUnsmithing
+        // UnSmithing
         toggleAllowUnsmithing = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.allow_unsmithing_", allowUnsmithing)), btn -> {
             allowUnsmithing = !allowUnsmithing;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.allow_unsmithing_", allowUnsmithing)));
-        }).bounds(x, (int) (baseY + 220 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(toggleAllowUnsmithing);
+        y += 50;
 
+        // Damaged
         toggleAllowDamaged = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.allow_damaged_", allowDamagedItems)), btn -> {
             allowDamagedItems = !allowDamagedItems;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.allow_damaged_", allowDamagedItems)));
-        }).bounds(x, (int) (baseY + 245 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(toggleAllowDamaged);
+        y += 25;
 
-        // Toggle for preventModdedIngredientsFromVanillaItems
+        toggleAllowDamagedNonRepairable = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.allow_damaged_non_repairable_", allowDamagedNonRepairable)), btn -> {
+            allowDamagedNonRepairable = !allowDamagedNonRepairable;
+            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.allow_damaged_non_repairable_", allowDamagedNonRepairable)));
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(toggleAllowDamagedNonRepairable);
+        y += 25;
+
+        minimumDurabilityInput = new EditBox(this.font, x, (int) (y - scrollAmount), widgetWidth, 20, Component.empty());
+        minimumDurabilityInput.setValue(Double.toString(minimumDurability));
+        this.addRenderableWidget(minimumDurabilityInput);
+        y += 50;
+
+        // Modded Ingredients
         togglePreventModdedIngredientsFromVanillaItems = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.prevent_modded_ingredients_from_vanilla_items_", preventModdedIngredientsFromVanillaItems)), btn -> {
             preventModdedIngredientsFromVanillaItems = !preventModdedIngredientsFromVanillaItems;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.prevent_modded_ingredients_from_vanilla_items_", preventModdedIngredientsFromVanillaItems)));
-        }).bounds(x, (int) (baseY + 270 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(togglePreventModdedIngredientsFromVanillaItems);
+        y += 25;
 
-        // Restricted Mod input box
         String joinedMod = String.join("\n", restrictedModIngredients);
-        restrictedModInput = MultiLineEditBox.builder().setX(x).setY((int) (baseY + 295 - scrollAmount)).build(font, widgetWidth, 90, Component.translatable("screen.uncrafteverything.blank"));
+        restrictedModInput = MultiLineEditBox.builder().setX(x).setY((int) (y - scrollAmount)).build(font, widgetWidth, 90, Component.translatable("screen.uncrafteverything.blank"));
         restrictedModInput.setValue(joinedMod);
         this.addRenderableWidget(restrictedModInput);
+        y += 120;
 
+        // Enchanted
+        toggleOutputEnchantedBook = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.output_enchanted_book_", outputEnchantedBook)), btn -> {
+            outputEnchantedBook = !outputEnchantedBook;
+            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.output_enchanted_book_", outputEnchantedBook)));
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(toggleOutputEnchantedBook);
+        y += 25;
+
+        toggleEnchantedBtn = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.allow_enchanted_", allowEnchantedItems)), btn -> {
+            allowEnchantedItems = !allowEnchantedItems;
+            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.allow_enchanted_", allowEnchantedItems)));
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(toggleEnchantedBtn);
+        y += 50;
+
+        // RecipeSelectionOrder
+        togglePrioritizeVanillaIngredientRecipe = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.prioritize_", prioritizeVanillaIngredientRecipe)), btn -> {
+            prioritizeVanillaIngredientRecipe = !prioritizeVanillaIngredientRecipe;
+            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.prioritize_", prioritizeVanillaIngredientRecipe)));
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(togglePrioritizeVanillaIngredientRecipe);
+        y += 50;
+
+        // FTB Quest Progression
         toggleEnableProgression = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.enable_progression_", enableProgression)), btn -> {
             enableProgression = !enableProgression;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.enable_progression_", enableProgression)));
-        }).bounds(x, (int) (baseY + 390 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(toggleEnableProgression);
+        y += 25;
 
         toggleOnlyAllowDefinedProgression = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.only_allow_defined_progression_", onlyAllowDefinedProgression)), btn -> {
             onlyAllowDefinedProgression = !onlyAllowDefinedProgression;
             btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.only_allow_defined_progression_", onlyAllowDefinedProgression)));
-        }).bounds(x, (int) (baseY + 415 - scrollAmount), widgetWidth, 20).build();
+        }).bounds(x, (int) (y - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(toggleOnlyAllowDefinedProgression);
 
-        // Toggle for toggleOutputEnchantedBook
-        toggleOutputEnchantedBook = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.output_enchanted_book_", outputEnchantedBook)), btn -> {
-            outputEnchantedBook = !outputEnchantedBook;
-            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.output_enchanted_book_", outputEnchantedBook)));
-        }).bounds(x, (int) (baseY + 440 - scrollAmount), widgetWidth, 20).build();
-        this.addRenderableWidget(toggleOutputEnchantedBook);
-
-        // Toggle for allowPrioritizeVanillaIngredientRecipe
-        togglePrioritizeVanillaIngredientRecipe = Button.builder(Component.translatable(getLabel("screen.uncrafteverything.config.prioritize_", prioritizeVanillaIngredientRecipe)), btn -> {
-            prioritizeVanillaIngredientRecipe = !prioritizeVanillaIngredientRecipe;
-            btn.setMessage(Component.translatable(getLabel("screen.uncrafteverything.config.prioritize_", prioritizeVanillaIngredientRecipe)));
-        }).bounds(x, (int) (baseY + 465 - scrollAmount), widgetWidth, 20).build();
-        this.addRenderableWidget(togglePrioritizeVanillaIngredientRecipe);
-
-        // Save button (always at bottom)
         saveButton = Button.builder(Component.translatable("screen.uncrafteverything.save"), this::pressSaveButton).bounds(this.width / 2 - 100, (this.height - 45) + 15, 200, 20).build();
         this.addRenderableWidget(saveButton);
     }
@@ -187,53 +211,76 @@ public class UEConfigScreen extends AbstractScrollableScreen {
         // Render labels with scroll offset
         int x = 10;
         int textWidth = this.width / 2 - 10;
-        int baseY = 30;
 
-        pGuiGraphics.text(this.font, Component.translatable("screen.uncrafteverything.config.restriction_type_label"), x, (int) (baseY - scrollAmount + (this.font.lineHeight / 2d) + 2), 0xFFFFFFFF);
+        pGuiGraphics.text(this.font, Component.translatable("screen.uncrafteverything.config.restriction_type_label"), x, (int) (restrictionTypeButton.getY() + (this.font.lineHeight / 2d) + 2), 0xFFFFFFFF);
 
         Component format = Component.translatable("screen.uncrafteverything.config.restricted_item_label");
-        pGuiGraphics.textWithWordWrap(this.font, format, x, (int) (baseY + 25 - scrollAmount + (this.font.lineHeight / 2d) + 20), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, format, x, (int) (restrictionsInput.getY() + (this.font.lineHeight / 2d) + 20), textWidth, 0xFFFFFFFF);
 
         // Format help text
         pGuiGraphics.pose().pushMatrix();
         pGuiGraphics.pose().scale(0.65f, 0.65f);
-        pGuiGraphics.pose().translate(x * 1.55f, (float) (((baseY + 25f - scrollAmount) * 1.6f) + this.font.wordWrapHeight(format, textWidth) * 2f - (this.font.lineHeight * 0.65f) + 40f));
+        pGuiGraphics.pose().translate(x * 1.55f, (float) ((restrictionsInput.getY() * 1.6f) + this.font.wordWrapHeight(format, textWidth) * 2f - (this.font.lineHeight * 0.65f) + 40f));
         pGuiGraphics.textWithWordWrap(this.font, Component.translatable("screen.uncrafteverything.config.format_label"), 0, 0, (int) (textWidth * 1.5), 0xFFAAAAAA);
         pGuiGraphics.pose().popMatrix();
 
         Component restrictAmbiguouslyCraftedItems = Component.translatable("screen.uncrafteverything.config.restrict_mbiguously_label");
-        pGuiGraphics.textWithWordWrap(this.font, restrictAmbiguouslyCraftedItems, x, (int) (baseY + 120 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(restrictAmbiguouslyCraftedItems, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, restrictAmbiguouslyCraftedItems, x, (int) (toggleRestrictAmbiguouslyCraftedItems.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(restrictAmbiguouslyCraftedItems, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component allowEnchantedItem = Component.translatable("screen.uncrafteverything.config.allow_enchanted_label");
-        pGuiGraphics.textWithWordWrap(this.font, allowEnchantedItem, x, (int) (baseY + 145 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowEnchantedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, allowEnchantedItem, x, (int) (toggleEnchantedBtn.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowEnchantedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
-        pGuiGraphics.textWithWordWrap(this.font, Component.translatable("screen.uncrafteverything.config.exp_type_label"), x, (int) (baseY + 170 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, Component.translatable("screen.uncrafteverything.config.exp_type_label"), x, (int) (toggleExpTypeBtn.getY() + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component expRequired = Component.translatable("screen.uncrafteverything.config.exp_required_label");
-        pGuiGraphics.textWithWordWrap(this.font, expRequired, x, (int) (baseY + 195 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(expRequired, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, expRequired, x, (int) (experienceInput.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(expRequired, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
-        pGuiGraphics.textWithWordWrap(this.font, Component.translatable("screen.uncrafteverything.config.allow_unsmithing_label"), x, (int) (baseY + 220 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, Component.translatable("screen.uncrafteverything.config.allow_unsmithing_label"), x, (int) (toggleAllowUnsmithing.getY() + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component allowDamagedItem = Component.translatable("screen.uncrafteverything.config.allow_damaged_label");
-        pGuiGraphics.textWithWordWrap(this.font, allowDamagedItem, x, (int) (baseY + 245 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowDamagedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, allowDamagedItem, x, (int) (toggleAllowDamaged.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowDamagedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+
+        Component allowDamagedNonRepairableItem = Component.translatable("screen.uncrafteverything.config.allow_damaged_non_repairable_label");
+        pGuiGraphics.textWithWordWrap(this.font, allowDamagedNonRepairableItem, x, (int) (toggleAllowDamagedNonRepairable.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowDamagedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+
+        Component minimumDurability = Component.translatable("screen.uncrafteverything.config.minimum_durability_label");
+        pGuiGraphics.textWithWordWrap(this.font, minimumDurability, x, (int) (minimumDurabilityInput.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowDamagedItem, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component preventModded = Component.translatable("screen.uncrafteverything.config.prevent_modded_ingredients_from_vanilla_items_label");
-        pGuiGraphics.textWithWordWrap(this.font, preventModded, x, (int) (baseY + 270 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(preventModded, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, preventModded, x, (int) (togglePreventModdedIngredientsFromVanillaItems.getY() + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(preventModded, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component restrictedMod = Component.translatable("screen.uncrafteverything.config.prevent_modid");
-        pGuiGraphics.textWithWordWrap(this.font, restrictedMod, x, (int) (baseY + 329 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(restrictedMod, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, restrictedMod, x, (int) (restrictedModInput.getY() + 34 + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(restrictedMod, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component enableProgression = Component.translatable("screen.uncrafteverything.config.enable_progression");
-        pGuiGraphics.textWithWordWrap(this.font, enableProgression, x, (int) (baseY + 392 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(enableProgression, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, enableProgression, x, (int) (toggleEnableProgression.getY() + 2 + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(enableProgression, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component onlyAllowDefined = Component.translatable("screen.uncrafteverything.config.only_allow_defined_progression");
-        pGuiGraphics.textWithWordWrap(this.font, onlyAllowDefined, x, (int) (baseY + 417 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(onlyAllowDefined, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, onlyAllowDefined, x, (int) (toggleOnlyAllowDefinedProgression.getY() + 2 + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(onlyAllowDefined, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component allowEnchantedBook = Component.translatable("screen.uncrafteverything.config.output_enchanted_book");
-        pGuiGraphics.textWithWordWrap(this.font, allowEnchantedBook, x, (int) (baseY + 443 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowEnchantedBook, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, allowEnchantedBook, x, (int) (toggleOutputEnchantedBook.getY() + 3 + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(allowEnchantedBook, textWidth) / 4d), textWidth, 0xFFFFFFFF);
 
         Component prioritizeVanillaIngredientRecipe = Component.translatable("screen.uncrafteverything.config.prioritize");
-        pGuiGraphics.textWithWordWrap(this.font, prioritizeVanillaIngredientRecipe, x, (int) (baseY + 468 - scrollAmount + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(prioritizeVanillaIngredientRecipe, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.textWithWordWrap(this.font, prioritizeVanillaIngredientRecipe, x, (int) (togglePrioritizeVanillaIngredientRecipe.getY() + 3 + (this.font.lineHeight / 2d) + 1 - this.font.wordWrapHeight(prioritizeVanillaIngredientRecipe, textWidth) / 4d), textWidth, 0xFFFFFFFF);
+
+        Component restrictions = Component.translatable("screen.uncrafteverything.config.restrictions").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component experience = Component.translatable("screen.uncrafteverything.config.experience").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component unSmithing = Component.translatable("screen.uncrafteverything.config.unsmithing").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component damaged = Component.translatable("screen.uncrafteverything.config.damaged").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component moddedIngredients = Component.translatable("screen.uncrafteverything.config.modded_ingredients").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component enchanted = Component.translatable("screen.uncrafteverything.config.enchanted").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component recipeSelectionOrder = Component.translatable("screen.uncrafteverything.config.recipe_selection_order").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+        Component ftbQuestProgression = Component.translatable("screen.uncrafteverything.config.ftb_quest_progression").withStyle(ChatFormatting.UNDERLINE, ChatFormatting.GOLD);
+
+        pGuiGraphics.centeredText(this.font, restrictions, this.width / 2, restrictionTypeButton.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, experience, this.width / 2, toggleExpTypeBtn.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, unSmithing, this.width / 2, toggleAllowUnsmithing.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, damaged, this.width / 2, toggleAllowDamaged.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, moddedIngredients, this.width / 2, togglePreventModdedIngredientsFromVanillaItems.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, enchanted, this.width / 2, toggleOutputEnchantedBook.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, recipeSelectionOrder, this.width / 2, togglePrioritizeVanillaIngredientRecipe.getY() - 25 + (font.lineHeight / 2), -1);
+        pGuiGraphics.centeredText(this.font, ftbQuestProgression, this.width / 2, toggleEnableProgression.getY() - 25 + (font.lineHeight / 2), -1);
 
         pGuiGraphics.disableScissor();
 
@@ -321,8 +368,9 @@ public class UEConfigScreen extends AbstractScrollableScreen {
         restrictions = Arrays.stream(restrictionsInput.getValue().split("\n")).map(String::trim).filter(s -> !s.isEmpty()).toList();
         experience = Integer.parseInt(experienceInput.getValue());
         restrictedModIngredients = Arrays.stream(restrictedModInput.getValue().split("\n")).map(String::trim).filter(s -> !s.isEmpty()).toList();
+        minimumDurability = Double.parseDouble(minimumDurabilityInput.getValue());
 
-        ServerBoundUEConfigPayload configPayload = new ServerBoundUEConfigPayload(restrictionType, restrictions, allowEnchantedItems, experienceType, experience, allowUnsmithing, allowDamagedItems, preventModdedIngredientsFromVanillaItems, restrictedModIngredients, enableProgression, onlyAllowDefinedProgression, outputEnchantedBook, prioritizeVanillaIngredientRecipe, restrictAmbiguouslyCraftedItems);
+        ServerBoundUEConfigPayload configPayload = new ServerBoundUEConfigPayload(restrictionType, restrictions, allowEnchantedItems, experienceType, experience, allowUnsmithing, allowDamagedItems, preventModdedIngredientsFromVanillaItems, restrictedModIngredients, enableProgression, onlyAllowDefinedProgression, outputEnchantedBook, prioritizeVanillaIngredientRecipe, restrictAmbiguouslyCraftedItems, allowDamagedNonRepairable, minimumDurability);
         Services.NETWORK.sendToServer(configPayload);
         Services.NETWORK.sendToServer(new ServerBoundRequestConfigPayload());
         onClose();
@@ -380,7 +428,7 @@ public class UEConfigScreen extends AbstractScrollableScreen {
             renderWrappedTooltip(guiGraphics, tooltip, mouseX, mouseY);
         }
 
-        if (toggleEnchantmentTypeBtn.isHovered()){
+        if (toggleExpTypeBtn.isHovered()){
             List<Component> tooltip = List.of(
                     title("tooltip.uncrafteverything.config.toggle_experience_type"),
                     valueInfo("tooltip.uncrafteverything.config.point", "tooltip.uncrafteverything.config.point_info"),
@@ -478,6 +526,24 @@ public class UEConfigScreen extends AbstractScrollableScreen {
                     valueInfo("tooltip.uncrafteverything.config.yes", "tooltip.uncrafteverything.config.toggle_prioritize_yes"),
                     Component.empty(),
                     valueInfo("tooltip.uncrafteverything.config.no", "tooltip.uncrafteverything.config.toggle_prioritize_no")
+            );
+            renderWrappedTooltip(guiGraphics, tooltip, mouseX, mouseY);
+        }
+
+        if (toggleAllowDamagedNonRepairable.isHovered()){
+            List<Component> tooltip = List.of(
+                    title("tooltip.uncrafteverything.config.toggle_allow_damaged_non_repairable_items"),
+                    valueInfo("tooltip.uncrafteverything.config.yes", "tooltip.uncrafteverything.config.toggle_allow_damaged_non_repairable_items_yes"),
+                    Component.empty(),
+                    valueInfo("tooltip.uncrafteverything.config.no", "tooltip.uncrafteverything.config.toggle_allow_damaged_non_repairable_items_no")
+            );
+            renderWrappedTooltip(guiGraphics, tooltip, mouseX, mouseY);
+        }
+
+        if (minimumDurabilityInput.isHovered()){
+            List<Component> tooltip = List.of(
+                    title("tooltip.uncrafteverything.config.minimum_durability"),
+                    description("tooltip.uncrafteverything.config.minimum_durability_description")
             );
             renderWrappedTooltip(guiGraphics, tooltip, mouseX, mouseY);
         }