+++ neoforge/src/main/java/com/coolerpromc/uncrafteverything/event/UEModBusEvent.java
@@ -2,11 +2,9 @@ package com.coolerpromc.uncrafteverything.event;
 
 import com.coolerpromc.uncrafteverything.Constants;
 import com.coolerpromc.uncrafteverything.blockentity.UEBlockEntities;
-import com.coolerpromc.uncrafteverything.config.NeoForgeUncraftEverythingClientConfig;
 import net.minecraft.core.Direction;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.fml.event.config.ModConfigEvent;
 import net.neoforged.neoforge.capabilities.Capabilities;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
 import net.neoforged.neoforge.transfer.item.WorldlyContainerWrapper;
@@ -21,19 +19,5 @@ public class UEModBusEvent {
             }
             return new WorldlyContainerWrapper(blockEntity.getInputHandler(), null);
         });
-    }
-
-    @SubscribeEvent
-    public static void onLoad(final ModConfigEvent.Loading event) {
-        if (event.getConfig().getSpec() == NeoForgeUncraftEverythingClientConfig.CONFIG_SPEC) {
-            NeoForgeUncraftEverythingClientConfig.CONFIG.onChanged();
-        }
-    }
-
-    @SubscribeEvent
-    public static void onReload(final ModConfigEvent.Reloading event) {
-        if (event.getConfig().getSpec() == NeoForgeUncraftEverythingClientConfig.CONFIG_SPEC) {
-            NeoForgeUncraftEverythingClientConfig.CONFIG.onChanged();
-        }
     }
 }