+++ neoforge/src/main/java/com/coolerpromc/uncrafteverything/platform/NeoForgePlatformHelper.java
@@ -2,6 +2,7 @@ package com.coolerpromc.uncrafteverything.platform;
 
 import com.coolerpromc.uncrafteverything.platform.services.IPlatformHelper;
 import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.FMLEnvironment;
 import net.neoforged.fml.loading.FMLLoader;
 import net.neoforged.fml.loading.FMLPaths;
 
@@ -30,5 +31,10 @@ public class NeoForgePlatformHelper implements IPlatformHelper {
     @Override
     public Path configPath() {
         return FMLPaths.CONFIGDIR.get();
+    }
+
+    @Override
+    public boolean isClient() {
+        return FMLEnvironment.getDist().isClient();
     }
 }