+++ changelog.md
@@ -1,4 +1,6 @@
-# 26.1.2.2
+# 26.1.2.3
 ### Changes
-- Added new config field `Restrict Ambiguously Crafted Items` to common config
-- Recipe that has ItemTags ingredient can be disabled by setting `Restrict Ambiguously Crafted Items` to `false`
+- Refactor config system, now use my own config lib (Most config will reset to default value)
+- Added new config field `allowDamagedNonRepairable` and `minimumDurability` to common config
+- Updated Uncraft Everything Config Screen
+- Re-implemented FTB Quest Compat